+++ ChangeLog.md
@@ -1,6 +1,7 @@
 #### v.4.0.6 (TBD)
 * Update to DICOM Standard 2020b.
 * Handle N-Event requests in DicomClient (required for synchronous storage commitment) (#1001)
+* Bug fix: Fix DICOM Tag conversion for private tags for element numbers xxff (#1059)
 
 #### v.4.0.5 (5/18/2020)
 * Bug fix: DicomTags of ValueRepresentation LT have not been validated.